equipment index Fixes #93
