+++ src/commands/inspect/inspect.ts
@@ -10,7 +10,7 @@ import {
 } from '@adventure-bot/character'
 import { actionEmbed } from '@adventure-bot/commands/inspect/actionEmbed'
 import { getHook } from '@adventure-bot/commands/inspect/getHook'
-import { itemEmbed } from '@adventure-bot/equipment/itemEmbed'
+import { itemEmbed } from '@adventure-bot/equipment'
 import { questEmbed } from '@adventure-bot/quest'
 import { statusEffectEmbed } from '@adventure-bot/statusEffects/statusEffectEmbed'
