+++ src/encounters/shop/sellItemPrompt.ts
@@ -15,8 +15,8 @@ import {
 import { getSaleRate } from '@adventure-bot/encounters/shop/getSaleRate'
 import { sellList } from '@adventure-bot/encounters/shop/sellList'
 import { sellValue } from '@adventure-bot/encounters/shop/sellValue'
-import { goldValue } from '@adventure-bot/equipment/goldValue'
-import { itemEmbed } from '@adventure-bot/equipment/itemEmbed'
+import { goldValue } from '@adventure-bot/equipment'
+import { itemEmbed } from '@adventure-bot/equipment'
 import store from '@adventure-bot/store'
 import { itemSold } from '@adventure-bot/store/slices/characters'
