+++ src/encounters/shop/sellValue.ts
@@ -1,5 +1,5 @@
 import { getSaleRate } from '@adventure-bot/encounters/shop/getSaleRate'
-import { Item } from '@adventure-bot/equipment/Item'
+import { Item } from '@adventure-bot/equipment'
 
 export function sellValue(item: Item): number {
   return Math.round(getSaleRate() * item.goldValue)
