+++ src/encounters/shop/shop.ts
@@ -10,10 +10,10 @@ import { times } from 'remeda'
 import { getCharacterUpdate, getUserCharacter } from '@adventure-bot/character'
 import { buyItemPrompt } from '@adventure-bot/encounters/shop/buyItemPrompt'
 import { sellItemPrompt } from '@adventure-bot/encounters/shop/sellItemPrompt'
-import { goldValue } from '@adventure-bot/equipment/goldValue'
-import { itemEmbed } from '@adventure-bot/equipment/itemEmbed'
-import { heavyCrown } from '@adventure-bot/equipment/items'
-import { randomShopItem } from '@adventure-bot/equipment/randomShopItem'
+import { goldValue } from '@adventure-bot/equipment'
+import { itemEmbed } from '@adventure-bot/equipment'
+import { heavyCrown } from '@adventure-bot/equipment'
+import { randomShopItem } from '@adventure-bot/equipment'
 import store from '@adventure-bot/store'
 import { selectIsHeavyCrownInPlay } from '@adventure-bot/store/selectors'
 import { asset } from '@adventure-bot/utils'
