+++ src/equipment/equipItemPrompt.ts
@@ -6,8 +6,8 @@ import {
 } from 'discord.js'
 
 import inspect from '@adventure-bot/commands/inspect/inspect'
-import { Item } from '@adventure-bot/equipment/Item'
-import { itemEmbed } from '@adventure-bot/equipment/itemEmbed'
+import { Item } from '@adventure-bot/equipment'
+import { itemEmbed } from '@adventure-bot/equipment'
 import store from '@adventure-bot/store'
 import { itemEquipped } from '@adventure-bot/store/slices/characters'
