+++ src/equipment/items/amuletOfAttack.ts
@@ -1,6 +1,6 @@
 import { randomUUID } from 'crypto'
 
-import { Amulet } from '@adventure-bot/equipment/equipment'
+import { Amulet } from '@adventure-bot/equipment'
 
 export const amuletOfAttack = (): Amulet => ({
   id: randomUUID(),
