+++ src/equipment/usableInventory.ts
@@ -1,5 +1,5 @@
 import { Character } from '@adventure-bot/character'
-import { Usable, isUsable } from '@adventure-bot/equipment/equipment'
+import { Usable, isUsable } from '@adventure-bot/equipment'
 
 export function usableInventory(character: Character): Usable[] {
   return character.inventory.filter(isUsable)
