+++ src/equipment/useInventoryItemPrompt.ts
@@ -10,9 +10,9 @@ import { clamp } from 'remeda'
 import { Emoji } from '@adventure-bot/Emoji'
 import { Manifest } from '@adventure-bot/asset-manifest'
 import { getUserCharacter, hpBarField } from '@adventure-bot/character'
-import { isPotion } from '@adventure-bot/equipment/equipment'
-import { itemSelect } from '@adventure-bot/equipment/itemSelect'
-import { usableInventory } from '@adventure-bot/equipment/usableInventory'
+import { isPotion } from '@adventure-bot/equipment'
+import { itemSelect } from '@adventure-bot/equipment'
+import { usableInventory } from '@adventure-bot/equipment'
 import { createEffect } from '@adventure-bot/statusEffects'
 import { statusEffectEmbed } from '@adventure-bot/statusEffects/statusEffectEmbed'
 import { EffectTemplate } from '@adventure-bot/statusEffects/templates'
