+++ src/store/slices/crown.ts
@@ -1,6 +1,6 @@
 import { createSlice } from '@reduxjs/toolkit'
 
-import { heavyCrownId } from '@adventure-bot/equipment/items/heavyCrown'
+import { heavyCrownId } from '@adventure-bot/equipment'
 import {
   backdateCrown,
   itemReceived,
